requirements.txt
torch==2.0.1
torchvision==0.15.2
streamlit==1.23.1
Pillow==9.4.0
supabase==2.5.1